+++ R/check-poly-formula.R
@@ -6,33 +6,57 @@
 #'
 #' @param formula A model formula in \code{x.name}.
 #' @param x.name character The name of the explanatory variable in the formula.
-#' @param warning.text character string.
+#' @param
+#'   warn.incr.poly.text,warn.transf.lhs.txt,warn.transf.rhs.txt,warn.as.is.txt,warn.poly.raw.txt,stop.pow.poly.text
+#'   character Text for warnings and errors.
+#' @param check.transf.rhs,check.transf.lhs logical flag enabling test for
+#'   transformation of variables.
 #'
-#' @details This validation check could fail to validate some valid formulas as
-#'   it is difficult to test, or even list all possible variations of valid
-#'   formulas. Consequently, this function triggers a warning in case of
-#'   failure, not an error. Furthermore, the statistics only fail to build the
-#'   correct equation label, but in most cases other output is still usable with
-#'   models that are not strictly polynomials.
+#' @details The assumption is that this function will be called from within a
+#'   ggplot2 compatible layer function, and that model formulas will always have
+#'   a single explanatory variable, variables will be \code{x} and \code{y}. Its
+#'   behaviour is undefined or erroneous in other cases.
+#'
+#'   This validation check could return a false positive or a false negative
+#'   results with some formulas as it is difficult to test, or even list all
+#'   possible variations of supported vs. unsupported formulas. This makes
+#'   testing difficult. In addition, many valid model formulas that can be
+#'   succesfully fitted, are not correctly converted into character labels.
+#'   Thus, this function triggers a warning in case of failure, not an error,
+#'   and returns a logic value. If this value is \code{FALSE}, the statistics in
+#'   'ggpmisc' skip the generation of an equation label, setting it to
+#'   \code{NA}. However, if the formula is accepted by the model fit function,
+#'   other labels and the numeric estimates of the fitted coefficients remain
+#'   usable. The stats can be used also with models that are not polynomials or
+#'   containing transformations.
 #'
 #'   Model formulas with and without an intercept term are accepted as valid, as
-#'   \code{+0}, \code{-1} and \code{+1} are accepted. If a single power term is
-#'   included, it is taken as a transformation and any power is accepted. If two
-#'   or more terms are powers, they are expected in increasing order with no
-#'   missing intermediate terms. If \code{poly()} is used in the model formula,
-#'   a single term is expected.
+#'   \code{+0}, \code{-1} and \code{+1} are accepted. If a single \code{as.is}
+#'   power term is included or if arithmetic (\code{sqrt()}, \code{exp()},
+#'   \code{log()}), or trigonometric functions (\code{cos()}, \code{sin()},
+#'   \code{tan()}, etc.) are encountered a warning is issued about the need to
+#'   pass a matching argument to parameter \code{eq.x.rhs} of the statistic.
 #'
-#'   This function checks that all power terms defined using \code{^} are
-#'   protected with "as is" \code{I()}, as otherwise they are not powers but
-#'   instead part of the formula specification. It also checks that an argument
-#'   is passed to parameter \code{raw} of function \code{poly()} if present.
+#'   If two or more terms are \code{as.is} (\code{I( )} protected) powers
+#'   (\code{^}), they are expected to be in increasing order with no missing
+#'   intermediate power terms. If \code{poly()} is used in the model formula, a
+#'   single term is expected. When calling function \code{poly()},
+#'   \code{raw = TRUE} must be passed to obtain suitable estimates for the
+#'   fitted coefficients, and this is also checked.
+#'
+#'   When the formula \emph{rhs} contains more than one power term, all power
+#'   terms defined using \code{^} must be protected as \code{"as.is"}
+#'   \code{I()}, as otherwise they are not powers but instead part of the
+#'   formula specification.
 #'
 #'   If the warning text is \code{NULL} or \code{character(0)} no warning is
-#'   issued. The caller always receives a length-1 logical as return value.
+#'   issued, but the test is done. In contrast, \code{check.transf.rhs,check =
+#'   FALSE} and \code{transf.lhs = FALSE} skip these two tests. The caller
+#'   always receives a length-1 logical as returned value.
 #'
-#' @return A logical, TRUE if the formula describes an increasing polynomial,
-#'   and FALSE otherwise. As a side-effect a warning is triggered when
-#'   validation fails.
+#' @return A logical, TRUE if the formula describes an increasing polynomial
+#'   suitable for conversion into a text label, and FALSE otherwise. When
+#'   validation fails, warnings are issued describing the problem encountered.
 #'
 #' @export
 #'
@@ -67,27 +91,72 @@
 check_poly_formula <-
   function(formula,
            x.name = "x",
-           warning.text = "'formula' not an increasing polynomial: 'eq.label' is NA!") {
+           warn.incr.poly.text = "'formula' not an increasing polynomial: 'eq.label' set to NA!",
+           warn.transf.rhs.txt = "rhs includes transformations requiring an argument for 'eq.x.rhs': 'eq.label' set to NA!.",
+           warn.transf.lhs.txt = "lhs includes transformations requiring an argument for 'eq.with.lhs': 'eq.label' set to NA!.",
+           warn.as.is.txt = "Power (^) terms in model formula of a polynomial need to be protected by 'I()': 'eq.label' set to NA!.",
+           warn.poly.raw.txt = "'poly()' in model formula has to be passed 'raw = TRUE': 'eq.label' set to NA!",
+           stop.pow.poly.text = "Both 'poly()' and power (^) terms in model formula.",
+           check.transf.rhs = TRUE,
+           check.transf.lhs = TRUE) {
+
+  transf.ok <- TRUE # flag use below
+
+  # parsing
+  lhs <- as.character(formula)[2]
   rhs <- as.character(formula)[3]
-  rhs.terms <- unlist(strsplit(x = rhs, split = c("+", "*"), fixed = TRUE))
+
+  formula.terms <- terms(formula)
+  rhs.terms <- attr(formula.terms, "term.labels")
+  # replaced as it fails when + or * appear within a function call such as I()
+  # rhs.terms <- unlist(strsplit(x = rhs, split = c("+", "*"), fixed = TRUE))
   num.terms <- length(rhs.terms)
   x.terms <- grepl(x.name, rhs.terms)
   poly.in.terms <- grepl("poly *\\(", as.character(formula)[3L])
   power.terms  <- grepl("\\^ *", rhs.terms)
+  if (!any(power.terms) && grepl("\\^ *", rhs)) {
+    warning("Terms in 'formula' were dropped! Forgotten 'I()'?")
+    # even if accepted downstream this indicates an error
+    return(FALSE)
+  }
   raw.terms  <- grepl("raw *=", rhs.terms)
   as.is.terms <- grepl("I *\\(", rhs.terms)
 
+  # checks
   if (num.terms > 1L && poly.in.terms && sum(power.terms) != 0L) {
-    stop("Both 'poly()' and power (^) terms in model formula.")
+    stop(stop.pow.poly.text)
   }
   if (num.terms > 1L && !all(which(power.terms) %in% which(as.is.terms))) {
-    warning("Power (^) terms in model formula of a polynomial need to be protected by 'I()'.")
+    if (length(warn.as.is.txt)) {
+      warning(warn.as.is.txt)
+    }
     return(FALSE)
   }
-  if (poly.in.terms && !sum(raw.terms)) {
-    warning("'poly()' in model formula has to be passed 'raw = TRUE'")
+  if (check.transf.rhs &&
+      (num.terms == 1L && any(as.is.terms) ||
+      any(grepl("log|exp|sqrt|cos|sin|tan", rhs)))) {
+    if (length(warn.transf.rhs.txt)) {
+      warning(warn.transf.rhs.txt)
+    }
+    transf.ok <- FALSE
   }
-  if (sum(x.terms) == 0L || poly.in.terms && num.terms == 1L) {
+  if (check.transf.lhs &&
+      any(grepl("log|exp|sqrt|cos|sin|tan|[-^/*+]|[^I[:space:]]\\(.*[xy].*\\)", lhs))) {
+    if (length(warn.transf.lhs.txt)) {
+      warning(warn.transf.lhs.txt)
+    }
+    transf.ok <- FALSE
+  }
+  if (!transf.ok) {
+    # flag ensures that both lhs and rhs are always checked and warnings issued
+    return(FALSE)
+  }
+  if (length(warn.poly.raw.txt) &&
+      poly.in.terms && !sum(raw.terms)) {
+    warning(warn.poly.raw.txt)
+    return(FALSE)
+  }
+  if (sum(x.terms) <= 1L || poly.in.terms && num.terms == 1L) {
     polynomial <- TRUE
     increasing <- TRUE
   } else if (sum(power.terms) < 1L ||
@@ -96,24 +165,32 @@ check_poly_formula <-
              sum(power.terms) == num.terms - 2L &&
              sum(x.terms) == num.terms - 1L) {
     polynomial <- TRUE
-    if (sum(x.terms) == 1L || min(which(power.terms)) %in% 2L:3L) {
-      powers <- as.numeric(gsub(".*\\^([0-9]+).*", "\\1", rhs.terms[power.terms]))
-      increasing <- length(powers) <= 1L ||
-        !is.unsorted(powers, strictly = TRUE) &&
-        max(powers) == length(powers) + 1 # no missing terms
+    if (sum(x.terms) >= 1L || any(power.terms)) {
+      powers <- numeric(length(x.terms)) # filled with zeros
+      if (sum(power.terms)) {
+        hi.powers.chr <- character(length(x.terms))
+        hi.powers.chr[power.terms] <-
+          gsub(".*\\^([0-9]+).*", "\\1", rhs.terms[power.terms])
+        powers[power.terms] <- as.numeric(hi.powers.chr[power.terms])
+      }
+      if (!any(powers == 1, na.rm = TRUE)) {
+        power1.terms <- x.terms & !power.terms
+        powers[power1.terms] <- 1
+      }
+      increasing <- min(powers) == 1 &&
+        (length(powers) == 1 || (length(powers) >= 2 && all(diff(powers) == 1L)))
     } else {
       increasing <- FALSE
     }
   } else {
-    polynomial = FALSE
+    polynomial <- FALSE
   }
   if (!polynomial || !increasing) {
-    if (length(warning.text)) {
-      warning(warning.text)
+    if (length(warn.incr.poly.text)) {
+      warning(warn.incr.poly.text)
     }
     FALSE
   } else {
     TRUE
   }
 }
-